metodos serviciosAlquilerimpl

+++ src/main/java/edu/eci/cvds/sampleprj/dao/mybatis/mappers/ClienteMapper.java
@@ -12,7 +12,7 @@ import edu.eci.cvds.samples.entities.Cliente;
  */
 public interface ClienteMapper {
     
-    public Cliente consultarCliente(@Param("idcli") int id); 
+    public Cliente consultarCliente(@Param("idcli") long id); 
     
     /**
      * Registrar un nuevo item rentado asociado al cliente identificado
@@ -22,7 +22,7 @@ public interface ClienteMapper {
      * @param fechainicio
      * @param fechafin 
      */
-     public void agregarItemRentadoACliente(@Param("idir")int id,
+     public void agregarItemRentadoACliente(@Param("idir") int id,
             @Param("idit")int idit, 
             @Param("fiir")Date fechainicio,
             @Param("ffir")Date fechafin);
